Bugfix in trace, out of range memory access in trace Bugfix in readConsole, changed input type from uint16_t to uint64_t Debug and realese configuration selection

+++ eset2/Evm/Operation.cpp
@@ -43,7 +43,7 @@ namespace Evm {
 		}
 
 		void ConsoleReadOperation::execute(ThreadContext & thread) {
-			uint16_t input;
+			uint64_t input;
 			ios_base::fmtflags flags{ cin.flags() };
 			cin >> hex >> input;
 			cin.flags(flags);
